[main.cpp] Use BSMP hook function to signal Attenuator Thread
The attenuator thread was using a polling scheme to check if there was a
new value to proccess. This is not efficient as it inserts a fixed
blocking delay in the thread, causing possible loss of data if the
attenuator value is in the block time.
Use the BSMP hook function to signal the attenuator thread when a new
value is written.